src/main/java/kasutajaliidesedApi/rest/jackson/map/LocalDateDeserializer.java
package kasutajaliidesedApi.rest.jackson.map;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;
import org.joda.time.DateTime;
import org.joda.time.LocalDate;

import java.io.IOException;

/**
 * Converts JSON date string into LocalDate.
 *
 */
public class LocalDateDeserializer extends JsonDeserializer<LocalDate> {

    @Override
    public LocalDate deserialize(JsonParser jp, DeserializationContext ctxt) throws IOException {
        return new DateTime(jp.getText()).toLocalDate();
    }
}
